refactor: Migrate UhcSpawnLocation to separate file and remove ?.let
